+++ packages/app/src/index.ts
@@ -1,3 +1,4 @@
 export * from './runtime'
+export * from './oplog'
 export * from './http'
 export { buildServerDeps, startServer } from './server'